admin/requirements.txt
quart
quart-cors
PyJWT
loguru
pydantic
pydantic_settings
loguru
sqlalchemy
psycopg2-binary
requests
